requirements.txt
# SALLY-E Bot Dependencies

# Flask web framework
Flask>=2.3.3
Flask-Cors>=4.0.0

# Telegram bot
python-telegram-bot>=20.6

# Database
mysql-connector-python>=8.2.0

# Blockchain/Web3
web3>=6.0.0
eth-account>=0.8.0
eth-keys>=0.4.0
mnemonic>=0.20

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Optional: PDF generation for reports
reportlab>=4.0.0
